requirements.txt
# Core dependencies
flask==3.0.2
werkzeug==3.0.1

# XML processing alternatives
defusedxml==0.7.1
xmltodict==0.13.0

# Security and cryptography
cryptography==41.0.1

# HTTP requests
requests==2.31.0

# QR Code generation
qrcode==7.3
pillow==11.0.0

# Configuration
python-dotenv==1.0.0

# Server
gunicorn==20.1.0